requirements.txt
streamlit>=1.32.0
pvlib>=0.11.0
plotly>=5.18.0
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
scipy>=1.11.0
python-dotenv>=1.0.0
pytz>=2024.1
openai>=1.3.0
google-generativeai>=0.3.0
